[~] change project structure
