+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/domain/autoplay/AutoplayableViewModel.kt
@@ -1,5 +1,6 @@
-package band.effective.office.tv.core.ui.autoplay
+package band.effective.office.tv.domain.autoplay
 
+import band.effective.office.tv.domain.autoplay.model.AutoplayState
 import kotlinx.coroutines.flow.StateFlow
 
 interface AutoplayableViewModel {
