+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/domain/autoplay/model/ScreenDescription.kt
@@ -1,7 +1,7 @@
-package band.effective.office.tv.screen.autoplay
+package band.effective.office.tv.domain.autoplay.model
 
 import androidx.compose.runtime.Composable
-import band.effective.office.tv.core.ui.autoplay.AutoplayableViewModel
+import band.effective.office.tv.domain.autoplay.AutoplayableViewModel
 import band.effective.office.tv.screen.navigation.Screen
 
 data class ScreenDescription(
