+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/domain/autoplay/model/ScreenState.kt
@@ -1,0 +1,5 @@
+package band.effective.office.tv.domain.autoplay.model
+
+enum class ScreenState {
+    Data, Error, Load
+}
