+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/screen/photo/PhotoViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import band.effective.office.tv.BuildConfig
 import band.effective.office.tv.core.network.entity.Either
-import band.effective.office.tv.core.ui.autoplay.AutoplayableViewModel
+import band.effective.office.tv.domain.autoplay.AutoplayableViewModel
 import band.effective.office.tv.core.ui.screen_with_controls.TimerSlideShow
 import band.effective.office.tv.repository.synology.SynologyRepository
 import band.effective.office.tv.screen.photo.model.toUIModel
